consolidation of sgxcclib, also in tlcc
