+++ common/sgxcclib/sgx_attestation_type.h
