+++ ecc_enclave/sgxcclib/CMakeLists.txt
@@ -48,7 +48,6 @@ add_custom_command(TARGET libsgxcc
     POST_BUILD
     COMMAND mkdir -p ${CMAKE_BINARY_DIR}/include
     COMMAND cp ${CMAKE_CURRENT_SOURCE_DIR}/sgxcclib.h ${CMAKE_BINARY_DIR}/include
-    COMMAND cp ${CMAKE_CURRENT_SOURCE_DIR}/sgx_attestation_type.h ${CMAKE_BINARY_DIR}/include
     COMMAND cp ${COMMON_SOURCE_DIR}/types.h ${CMAKE_BINARY_DIR}/include
     COMMAND mkdir -p ${CMAKE_BINARY_DIR}/lib/
     COMMAND cp ${CMAKE_BINARY_DIR}/libsgxcc.so ${CMAKE_BINARY_DIR}/lib/
