+++ ecc_enclave/sgxcclib/sgxcclib.c
@@ -7,7 +7,6 @@
 
 #include "sgxcclib.h"
 #include "enclave_u.h"
-#include "sgx_attestation_type.h"
 
 #include <stdbool.h>
 #include <string.h>
@@ -19,48 +18,8 @@
 
 #include "sgx_urts.h"
 
-#define NRM "\x1B[0m"
-#define RED "\x1B[31m"
-#define CYN "\x1B[36m"
-
-//TODO: separate logging in sgxcclib
-
-// Prototypes of CGo functions implemented in ecc/enclave/enclave_stub.go
-// - logging
-extern void golog(const char* format, ...);
-
-#include <stdio.h>
-#include <stdarg.h>
-
-#define BUF_SIZE 1024
-#define LARGE_BUF_SIZE (BUF_SIZE*2)
-void LOG_ERROR(const char* fmt, ...) {
-    //create message
-    char msg[BUF_SIZE];
-    va_list ap;
-    va_start(ap, fmt);
-    vsnprintf(msg, BUF_SIZE, fmt, ap);
-    va_end(ap);
-    //color the message
-    char colored_msg[LARGE_BUF_SIZE];
-    snprintf(colored_msg, LARGE_BUF_SIZE, RED "ERROR: %s" NRM "\n", msg);
-    //dump message
-    golog(colored_msg);
-}
-
-void LOG_DEBUG(const char* fmt, ...) {
-    //create message
-    char msg[BUF_SIZE];
-    va_list ap;
-    va_start(ap, fmt);
-    vsnprintf(msg, BUF_SIZE, fmt, ap);
-    va_end(ap);
-    //color the message
-    char colored_msg[LARGE_BUF_SIZE];
-    snprintf(colored_msg, LARGE_BUF_SIZE, CYN "DEBUG: %s" NRM "\n", msg);
-    //dump message
-    golog(colored_msg);
-}
+extern void LOG_DEBUG(const char* fmt, ...);
+extern void LOG_ERROR(const char* fmt, ...);
 
 // - creator access
 extern void get_creator_name(
@@ -81,50 +40,6 @@ extern void get_state_by_partial_composite_key(const char* comp_key,
     void* ctx);
 extern void put_state(const char* key, uint8_t* val, uint32_t val_len, void* ctx);
 
-int sgxcc_get_quote_size(uint8_t *p_sig_rl, uint32_t sig_rl_size, uint32_t *p_quote_size)
-{
-    *p_quote_size = 0;
-    if(sig_rl_size > 0 && p_sig_rl == NULL) {
-        LOG_ERROR("Lib: Error: sigrlsize not zero but null sig_rl");
-        return SGX_ERROR_UNEXPECTED;
-    }
-    if(p_quote_size == NULL) {
-        LOG_ERROR("Lib: Error: pquotesize is null");
-        return SGX_ERROR_UNEXPECTED;
-    }
-
-    sgx_status_t ret = sgx_calc_quote_size(p_sig_rl, sig_rl_size, p_quote_size);
-    if(ret != SGX_SUCCESS) {
-        LOG_ERROR("Lib: Error: %d", ret);
-    }
-
-    return (int)ret;
-}
-
-int sgxcc_get_target_info(enclave_id_t eid, target_info_t* p_target_info)
-{
-    int ret = sgx_get_target_info(eid, (sgx_target_info_t*)p_target_info);
-    if (ret != SGX_SUCCESS)
-    {
-        LOG_ERROR("Lib: ERROR - sgx_get_target_info: %d", ret);
-    }
-
-    return ret;
-}
-
-int sgxcc_get_local_attestation_report(
-    enclave_id_t eid, target_info_t* target_info, report_t* report, ec256_public_t* pubkey)
-{
-    int enclave_ret = SGX_ERROR_UNEXPECTED;
-    int ret = ecall_create_report(eid, &enclave_ret, (sgx_target_info_t*)target_info,
-        (sgx_report_t*)report, (uint8_t*)pubkey);
-    if (ret != SGX_SUCCESS || enclave_ret != SGX_SUCCESS)
-    {
-        LOG_ERROR("Lib: ERROR - ecall_create_report: %d", ret);
-    }
-    return enclave_ret;
-}
-
 int sgxcc_bind(enclave_id_t eid, report_t* report, ec256_public_t* pubkey)
 {
     int enclave_ret = SGX_ERROR_UNEXPECTED;
@@ -132,68 +47,6 @@ int sgxcc_bind(enclave_id_t eid, report_t* report, ec256_public_t* pubkey)
     if (ret != SGX_SUCCESS || enclave_ret != SGX_SUCCESS)
     {
         LOG_ERROR("Lib: ERROR - ecall_bind_tlcc: %d", ret);
-    }
-    return enclave_ret;
-}
-
-int sgxcc_get_remote_attestation_report(
-    enclave_id_t eid, quote_t* quote, uint32_t quote_size, ec256_public_t* pubkey, spid_t* spid, uint8_t *p_sig_rl, uint32_t sig_rl_size)
-{
-    sgx_target_info_t qe_target_info = {0};
-    sgx_epid_group_id_t gid = {0};
-    sgx_report_t report;
-
-    int ret = sgx_init_quote(&qe_target_info, &gid);
-    if (ret != SGX_SUCCESS)
-    {
-        LOG_ERROR("Lib: ERROR - sgx_init_quote: %d", ret);
-        return ret;
-    }
-
-    // get report from enclave
-    int enclave_ret = SGX_ERROR_UNEXPECTED;
-    ret = ecall_create_report(eid, &enclave_ret, &qe_target_info, &report, (uint8_t*)pubkey);
-    if (ret != SGX_SUCCESS || enclave_ret != SGX_SUCCESS)
-    {
-        LOG_ERROR("Lib: ERROR - ecall_create_report: %d", ret);
-        return enclave_ret;
-    }
-
-    uint32_t required_quote_size = 0;
-    ret = sgxcc_get_quote_size(p_sig_rl, sig_rl_size, &required_quote_size);
-    if(ret != SGX_SUCCESS)
-    {
-        LOG_ERROR("Lib: ERROR - get quote size, %d", ret);
-        return ret;
-    }
-    if (quote_size < required_quote_size)
-    {
-        LOG_ERROR("Lib: ERROR - quote size too small. Required %u have %u", required_quote_size, quote_size);
-        return SGX_ERROR_OUT_OF_MEMORY;
-    }
-
-    ret = sgx_get_quote(&report, SGX_QUOTE_SIGN_TYPE,
-        (sgx_spid_t*)spid,  // spid
-        NULL,               // nonce
-        p_sig_rl,           // sig_rl
-        sig_rl_size,        // sig_rl_size
-        NULL,               // p_qe_report
-        (sgx_quote_t*)quote, quote_size);
-
-    if (ret != SGX_SUCCESS)
-    {
-        LOG_ERROR("Lib: ERROR - sgx_get_quote: %d", ret);
-    }
-    return ret;
-}
-
-int sgxcc_get_pk(enclave_id_t eid, ec256_public_t* pubkey)
-{
-    int enclave_ret = SGX_ERROR_UNEXPECTED;
-    int ret = ecall_get_pk(eid, &enclave_ret, (uint8_t*)pubkey);
-    if (ret != SGX_SUCCESS || enclave_ret != SGX_SUCCESS)
-    {
-        LOG_ERROR("Lib: ERROR - ecall_get_pk: %d", ret);
     }
     return enclave_ret;
 }
@@ -241,17 +94,6 @@ int sgxcc_invoke(enclave_id_t eid,
     }
 
     return enclave_ret;
-}
-
-int sgxcc_get_egid(unsigned int* p_egid) {
-    sgx_target_info_t target_info = {0};
-    int ret = sgx_init_quote(&target_info, (sgx_epid_group_id_t*)p_egid);
-    if (ret != SGX_SUCCESS)
-    {
-        LOG_ERROR("Lib: ERROR - sgx_get_egid: %d", ret);
-    }
-    LOG_DEBUG("EGID: %u", *p_egid);
-    return ret;
 }
 
 /* OCall functions */
