+++ ecc_enclave/sgxcclib/sgxcclib.h
@@ -14,15 +14,8 @@
 extern "C" {
 #endif
 
-int sgxcc_get_quote_size(uint8_t *p_sig_rl, uint32_t sig_rl_size, uint32_t *p_quote_size);
-
-int sgxcc_get_local_attestation_report(
-    enclave_id_t eid, target_info_t* target_info, report_t* report, ec256_public_t* pubkey);
-
 int sgxcc_get_remote_attestation_report(
     enclave_id_t eid, quote_t* quote, uint32_t quote_size, ec256_public_t* pubkey, spid_t* spid, uint8_t *p_sig_rl, uint32_t sig_rl_size);
-
-int sgxcc_get_target_info(enclave_id_t eid, target_info_t* target_info);
 
 int sgxcc_bind(enclave_id_t eid, report_t* report, ec256_public_t* pubkey);
 
@@ -45,10 +38,6 @@ int sgxcc_invoke(enclave_id_t eid,
     uint32_t* response_len_out,
     ec256_signature_t* signature,
     void* ctx);
-
-int sgxcc_get_pk(enclave_id_t eid, ec256_public_t* pubkey);
-
-int sgxcc_get_egid(unsigned int* p_egid);
 
 #ifdef __cplusplus
 }
